refactor(rome_fs): adjust some code that make it more clearer
